feat(task): Add yield_now #370
